network unix protocols support auto create directroy

+++ lib/network.c
@@ -652,7 +652,24 @@ int unix_create_service(int type, const char *pathname)
 
 	if (pathname && pathname[0])
 	{
-		unlink(pathname);
+		if (file_access_e(pathname))
+		{
+			unlink(pathname);
+		}
+		else
+		{
+			char dirname[PATH_MAX];
+
+			text_dirname_base(dirname, pathname);
+
+			ret = mkdir_hierarchy(dirname, 0777);
+			if (ret < 0)
+			{
+				pr_red_info("mkdir_hierarchy");
+				goto out_close_sockfd;
+			}
+		}
+
 		unix_sockaddr_init(&addr, pathname);
 		addrlen = sizeof(struct sockaddr_un);
 	}
@@ -666,11 +683,14 @@ int unix_create_service(int type, const char *pathname)
 	if (ret < 0)
 	{
 		print_error("bind");
-		close(sockfd);
-		return ret;
+		goto out_close_sockfd;
 	}
 
 	return sockfd;
+
+out_close_sockfd:
+	close(sockfd);
+	return ret;
 }
 
 int inet_create_tcp_service(u16 port)
@@ -2982,20 +3002,9 @@ int network_client_get_remote_ip(struct network_client *client, struct in_addr *
 
 int network_service_open(struct network_service *service, const struct network_url *url, int flags)
 {
-	int ret;
 	const struct network_protocol_desc *desc;
 
 	pd_bold_info("URL = %s", network_url_tostring(url, NULL, 0, NULL));
-
-	if (service->type == NETWORK_PROTOCOL_UNIX_TCP || service->type == NETWORK_PROTOCOL_UNIX_UDP)
-	{
-		ret = mkdir_hierarchy(CAVAN_NETWORK_TEMP_PATH, 0777);
-		if (ret < 0)
-		{
-			pr_red_info("mkdir_hierarchy %s", CAVAN_NETWORK_TEMP_PATH);
-			return ret;
-		}
-	}
 
 	desc = network_get_protocol_by_name(url->protocol);
 	if (desc == NULL)
